+++ UniTAS/Patcher/Models/UnitySafeWrappers/CursorLockMode.cs
@@ -1,0 +1,8 @@
+namespace UniTAS.Patcher.Models.UnitySafeWrappers;
+
+public enum CursorLockMode
+{
+    None,
+    Locked,
+    Confined
+}